Commit #155 and #157
