+++ scvi/inference/__init__.py
@@ -1,18 +1,18 @@
 from .inference import Inference
-from .classifier_inference import ClassifierInference
 from .variational_inference import (
     VariationalInference,
+    VariationalInferenceFish
+)
+from .classifier_inference import (
     AlternateSemiSupervisedVariationalInference,
     JointSemiSupervisedVariationalInference,
-    VariationalInferenceFish
+    ClassifierInference
 )
 from .experimental_inference import adversarial_wrapper, mmd_wrapper
 
 __all__ = ['Inference',
            'ClassifierInference',
            'VariationalInference',
-           'AlternateSemiSupervisedVariationalInference',
-           'JointSemiSupervisedVariationalInference',
            'VariationalInferenceFish',
            'adversarial_wrapper',
            'mmd_wrapper']
