--- scvi/metrics/early_stopping.py
@@ -1,62 +1,0 @@
-import numpy as np
-
-
-class EarlyStopping:
-    def __init__(self, patience=15, threshold=3, benchmark=False, mode="max", mode_save_state="max"):
-        self.benchmark = benchmark
-        self.patience = patience
-        self.threshold = threshold
-        self.epoch = 0
-        self.wait = 0
-        self.mode = mode
-        # We set the best to + inf because we're dealing with a loss we want to minimize
-        self.current_performance = np.inf
-        self.best_performance = np.inf
-        self.best_performance_state = np.inf
-        # If we want to maximize, we start at - inf
-        if self.mode == "max":
-            self.best_performance *= -1
-            self.current_performance *= -1
-        self.mode_save_state = mode_save_state
-        if self.mode_save_state == "max":
-            self.best_performance_state *= -1
-
-    def update(self, scalar):
-        self.epoch += 1
-        if self.benchmark or self.epoch < self.patience:
-            continue_training = True
-        elif self.wait >= self.patience:
-            continue_training = False
-        else:
-            # Shift
-            self.current_performance = scalar
-
-            # Compute improvement
-            if self.mode == "max":
-                improvement = self.current_performance - self.best_performance
-            elif self.mode == "min":
-                improvement = self.best_performance - self.current_performance
-
-            # updating best performance
-            if improvement > 0:
-                self.best_performance = self.current_performance
-
-            if improvement < self.threshold:
-                self.wait += 1
-            else:
-                self.wait = 0
-
-            continue_training = True
-        if not continue_training:
-            print("\nStopping early: no improvement of more than " + str(self.threshold) +
-                  " nats in " + str(self.patience) + " epochs")
-            print("If the early stopping criterion is too strong, "
-                  "please instantiate it with different parameters in the train method.")
-        return continue_training
-
-    def update_state(self, scalar):
-        improved = ((self.mode_save_state == "max" and scalar - self.best_performance_state > 0) or
-                    (self.mode_save_state == "min" and self.best_performance_state - scalar > 0))
-        if improved:
-            self.best_performance_state = scalar
-        return improved
