+++ scvi/models/classifier.py
@@ -4,7 +4,6 @@ from sklearn.linear_model import LogisticRegression
 from torch import nn as nn
 from torch.nn import Linear
 
-from scvi.metrics.clustering import get_latent_mean
 from scvi.models.modules import FCLayers
 
 
@@ -17,16 +16,3 @@ class Classifier(nn.Module):
 
     def forward(self, x):
         return self.classifier(x)
-
-
-class LinearLogRegClassifier(Linear):
-    def update_parameters(self, vae, data_loader_classification):
-        data_train, _, labels_train = get_latent_mean(vae, data_loader_classification)
-        log_reg = LogisticRegression()
-        log_reg.fit(data_train, labels_train)
-
-        self.weight = nn.Parameter(torch.from_numpy(log_reg.coef_).to(self.weight.device).type(self.weight.dtype))
-        self.bias = nn.Parameter(torch.from_numpy(log_reg.intercept_).to(self.bias.device).type(self.bias.dtype))
-
-    def forward(self, input):
-        return F.softmax(super(LinearLogRegClassifier, self).forward(input), dim=-1)
