+++ scvi/models/vae_fish.py
@@ -5,11 +5,11 @@ import torch
 import torch.nn.functional as F
 from torch.distributions import Normal, Poisson, kl_divergence as kl
 
-from scvi.metrics.log_likelihood import log_zinb_positive, log_nb_positive
-from scvi.models.modules import Encoder, DecoderSCVI
-from scvi.models.classifier import Classifier
-from scvi.models.utils import one_hot
 from scvi.models import VAE
+from scvi.models.classifier import Classifier
+from scvi.models.log_likelihood import log_zinb_positive, log_nb_positive
+from scvi.models.modules import Encoder, DecoderSCVI
+from scvi.models.utils import one_hot
 
 torch.backends.cudnn.benchmark = True
